requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cssbeautifier==1.15.1
Django==5.1.5
djlint==1.36.4
EditorConfig==0.17.0
jsbeautifier==1.15.1
json5==0.10.0
mysqlclient==2.2.7
passlib==1.7.4
pathspec==0.12.1
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
